CrashesListenerProvider class added for different flavours.
